requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
python-dotenv==1.0.0
qrcode==7.3
Pillow==9.5.0
pandas==2.2.0
requests==2.31.0
razorpay==1.4.2
gunicorn
